Extract interface localization helpers
